Fixed issue #37, [Enhancement] Refactor RAGStreamHandler to use LangGraph Pipeline
