+++ .env.example
@@ -1,34 +1,62 @@
-API_KEY=xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
-LLM_MODEL="qwen-turbo-2024-11-01"
-LLM_BASE_URL="https://dashscope.aliyuncs.com/compatible-mode/v1"
-
-HF_ENDPOINT=https://hf-mirror.com
-BGE_MODEL_PATH=BAAI/bge-m3
-
-OLLAMA_URL=
-OLLAMA_TOKEN=
-
-# Please fill in the same value as API_KEY for OPENAI_EMBEDDING_API_KEY
-OPENAI_EMBEDDING_API_KEY=xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
-OPENAI_EMBEDDING_BASE_URL="https://dashscope.aliyuncs.com/compatible-mode/v1/embeddings"
-OPENAI_EMBEDDING_MODEL=text-embedding-v3
-
-# Embedding model type
-# Options: default, local_model, ollama, openai_embedding
-# default: Use the default bge-m3 embedding model
-# local_model: Use a local embedding model
-# ollama: Use Ollama for embeddings
-# openai_embedding: Use OpenAI-compatible embedding API
-EMBEDDED_TYPE=default
 
 UI_LANG="zh"
+
+
+#######################################################################
+###################                               #####################
+###################          Model Setting        #####################
+###################                               #####################
+#######################################################################
+HF_ENDPOINT=https://hf-mirror.com
+
+### Chat model
+API_KEY=xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
+LLM_MODEL="qwen3-coder-plus"
+LLM_BASE_URL="https://dashscope.aliyuncs.com/compatible-mode/v1"
+
+# =============================================================================
+# Embedding Model Configuration
+# =============================================================================
+# EMBEDDED_TYPE: Embedding model type, available options:
+#   - default: Use built-in sentence-transformers/all-MiniLM-L6-v2 model (no additional config needed)
+#   - local_model: Use local embedding model (requires EMBEDDED_LLM_MODEL and EMBEDDED_LLM_BASE_URL)
+#   - ollama: Use Ollama embedding service (requires all three params below)
+#   - openai_embedding: Use OpenAI embedding API (requires all three params below)
+EMBEDDED_TYPE=default
+
+# Vector embedding dimension (must match your embedding model's output dimension)
+EMBEDDED_DIMENSION=384
+
+# EMBEDDED_API_KEY: API key for embedding service
+#   - Required for: ollama, openai_embedding
+#   - Not required for: default, local_model
+EMBEDDED_API_KEY=
+
+# EMBEDDED_LLM_MODEL: Embedding model name
+#   - For local_model: model name (e.g., BAAI/bge-m3)
+#   - For ollama: model name (e.g., nomic-embed-text)
+#   - For openai_embedding: model name (e.g., tongyi text-embedding-3-small)
+EMBEDDED_LLM_MODEL=
+
+# EMBEDDED_LLM_BASE_URL: Base URL or model path
+#   - For local_model: local model path (e.g., /path/to/model), if this is empty, it will be automatically downloaded
+#   - For ollama: Ollama server URL (e.g., http://localhost:11434)
+#   - For openai_embedding: OpenAI API base URL (e.g., https://api.openai.com/v1)
+EMBEDDED_LLM_BASE_URL=
+
+
+#######################################################################
+###################                               #####################
+###################          Database Setting     #####################
+###################                               #####################
+#######################################################################
 
 # Whether to reuse the current database
 # When set to true, will reuse existing database connection
 # When set to false, will start download dockers, at this time:
 #                   if DB_STORE is seekdb, DB_USER must be root
 #                   if DB_STORE is oceanbase, DB_USER must be root@test
-REUSE_CURRENT_DB=false
+REUSE_CURRENT_DB=true
 
 
 # Use what kind of docker, seekdb's docker or oceanbase-ce's docker
@@ -45,5 +73,21 @@ DB_PORT="2881"
 DB_USER="root"
 # If database use OceanBase, the DB_USER will contain tenant's name
 # DB_USER="root@test"
-DB_PASSWORD="Testob1234"
+DB_PASSWORD="root@test"
 DB_NAME="test"
+
+
+#######################################################################
+###################                               #####################
+###################         RAG Parser Setting    #####################
+###################                               #####################
+#######################################################################
+# Maximum chunk size for text splitting (in characters)
+MAX_CHUNK_SIZE=4096
+
+# Limit the number of documents to process (0 means no limit)
+LIMIT=0
+
+# Patterns to skip when processing documents (comma-separated, e.g., "*.log,*.tmp")
+SKIP_PATTERNS=""
+
