+++ pyproject.toml
@@ -14,6 +14,7 @@ dependencies = [
     "langchain-core>=1.0.0",
     "langchain-ollama>=0.2.0",
     "langchain-text-splitters>=1.0.0",
+    "langgraph==1.0.6",
     "streamlit>=1.38.0",
     "pyseekdb>=0.1.0",
     "langchain-oceanbase>=0.3.2",
