+++ src/common/logger.py
@@ -33,7 +33,7 @@ def setup_logger(name: Optional[str] = None, log_level: int = logging.INFO) -> l
     # Create formatter with standard format
     # Format: [MODULE:LINE - YYYY-MM-DD HH:MM:SS,mmm - LEVEL] - MESSAGE
     formatter = logging.Formatter(
-        fmt='[%(name)s:%(lineno)d-%(asctime)s-%(levelname)s] %(message)s',
+        fmt='[%(name)s:%(lineno)d %(asctime)s %(levelname)s] %(message)s',
         datefmt='%Y-%m-%d %H:%M:%S'
     )
